Revert "Revert "Squashed commit of the following:""
This reverts commit 3ea6a3fb0ac76fe58199b7de2b4311dab318f73c.

+++ src_c/scrap_qnx.c
@@ -72,7 +72,7 @@ pygame_scrap_lost(void)
 }
 
 int
-pygame_scrap_put(char *type, int srclen, char *src)
+pygame_scrap_put(char *type, Py_ssize_t srclen, char *src)
 {
     uint32_t format;
     int nulledlen = srclen + 1;
